+++ packages/iov-encoding/types/integers.d.ts
@@ -1,12 +1,18 @@
+/** Internal interface to ensure all integer types can be used equally */
 interface Integer {
   readonly toNumber: () => number;
   readonly toString: () => string;
 }
-export declare class Uint32 implements Integer {
+interface WithByteConverters {
+  readonly toBytesBigEndian: () => readonly number[];
+  readonly toBytesLittleEndian: () => readonly number[];
+}
+export declare class Uint32 implements Integer, WithByteConverters {
   static fromBigEndianBytes(bytes: ArrayLike<number>): Uint32;
   protected readonly data: number;
   constructor(input: number);
   toBytesBigEndian(): readonly number[];
+  toBytesLittleEndian(): readonly number[];
   toNumber(): number;
   toString(): string;
 }
@@ -24,7 +30,7 @@ export declare class Uint53 implements Integer {
   toNumber(): number;
   toString(): string;
 }
-export declare class Uint64 implements Integer {
+export declare class Uint64 implements Integer, WithByteConverters {
   static fromBytesBigEndian(bytes: ArrayLike<number>): Uint64;
   static fromString(str: string): Uint64;
   static fromNumber(input: number): Uint64;